refactor(prerender): move toPathname to utils.js and pass pre-fetched suggestions to syncSuggestions
